pyproject.toml
[project]
name = "beyond-pattern-matching"
version = "0.1.0"
description = "Towards true AI understanding"
authors = [
    {name = "Umut Eser",email = "umuteser@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "torch (<2.8.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
